Add question
Each node of the graph is a finite state machine.  From my understanding
of computation theory, it stands to reason that each of these machines
must have an input.  What is this input?

Are the inputs the entire graph?  Is each privilege just a transition
from state to state?  What is going on?