A bona-fide Python module!
Slightly reworked the filesystem.  See the Org file for details on the
why and how, but this (hopefully) simplifies distribution as well as
automated testing.

On that note, I separated out test cases into a separate directory.  The
command `nosetests` is now able to search for each test case and run it
to ensure that we get the expected output.

It should be noted that the first test case---the random graph
generator---uses margins of error in its assertions.  It doesn't make
sense to be able to predict the outcome.  Given the nature of this test
case, it takes a little while to complete.  (No longer than ten seconds
on a 2.26GHz Core 2 Duo.)