Use Travis hooks appropriately
before_install should be used to install dependencies.  Failures here
will not result in a failed build.

install is used to install the project, as it's now correctly doing.

test runs the test scripts.

I removed dependency on target check from target test since the two
are really pretty distinct and can be done as separate steps in the
Travis configuration.