Add bundling functionality
We're not using annotations from __future__ since CI is having real
trouble with installing Python 3.7.  This would help for the static
methods of `Bundle' that return a `Bundle' object -- the difficulty
being that `Bundle' is not yet defined when the function is read by
Python.  The workaround is to use strings that can be resolved later.